+++ dart/test/sentry_envelope_header_test.dart
@@ -1,5 +1,6 @@
 import 'package:sentry/sentry.dart';
 import 'package:sentry/src/sentry_envelope_header.dart';
+import 'package:sentry/src/utils.dart';
 import 'package:test/test.dart';
 
 import 'mocks.dart';
@@ -22,11 +23,13 @@ void main() {
         'trace_id': '${SentryId.newId()}',
         'public_key': '123',
       });
+      final timestamp = DateTime.utc(2019);
       final sut = SentryEnvelopeHeader(
         eventId,
         sdkVersion,
         dsn: fakeDsn,
         traceContext: context,
+        sentAt: timestamp,
       );
       final expextedSkd = sdkVersion.toJson();
       final expected = <String, dynamic>{
@@ -34,6 +37,7 @@ void main() {
         'sdk': expextedSkd,
         'trace': context.toJson(),
         'dsn': fakeDsn,
+        'sent_at': formatDateAsIso8601WithMillisPrecision(timestamp),
       };
       expect(sut.toJson(), expected);
     });